+++ pkg/webkit/script_cgo.go
@@ -4,6 +4,7 @@ package webkit
 
 /*
 #cgo pkg-config: webkitgtk-6.0 gtk4
+#cgo CFLAGS: -I/usr/include/webkitgtk-6.0
 #include <stdlib.h>
 #include <webkit/webkit.h>
 */